requirements.txt
Django==5.2.8
pymongo
python-dotenv
deepface
opencv-python
Pillow
dnspython
tensorflow
keras
gunicorn
